+++ src/components/SlideNavigation.js
