+++ src/components/UserProfile.js
@@ -5,12 +5,7 @@ import axiosClient from '../lib/axiosClient';
 import { useEffect } from 'react';
 import HelpButton from './ui/HelpButton';
 import appa from '../img/appa.png';
-import P1 from '../img/P1.jpg';
-import P2 from '../img/P2.jpg';
-import P3 from '../img/P3.jpg';
-import P7 from '../img/P7.jpg';
-import P5 from '../img/P5.jpg';
-import P6 from '../img/P6.jpg';
+import AvatarGallery from '../assets/AvatarGallery';
 
 const UserProfile = () => {
     const [name, setName] = useState("");
@@ -20,6 +15,7 @@ const UserProfile = () => {
     const [isEditing, setIsEditing] = useState(false);
     const [helpText, setHelpText] = useState("");
     const userId = document.cookie
+
         .split("; ")
         .find((row) => row.startsWith("user_id="))
         ?.split("=")[1];
@@ -125,46 +121,7 @@ const UserProfile = () => {
                 <div className="mt-10 w-fit mx-auto xl:mt-2 lg:mt-2 md:mt-0">
                     {isEditing ? (
                         <>
-                            <label className='block font-bold text-lg text-left'>Profile Pictures:</label>
-                            <div className="flex justify-items-center space-x-6">
-                                <img
-                                    src={P1}
-                                    alt="Avatar 1"
-                                    className={`w-16 h-16 rounded-full cursor-pointer transform hover:scale-125 transition duration-300 ${avatar === P1 ? 'border-2 border-oraange' : ''}`}
-                                    onClick={() => handleAvatarSelect(P1)}
-                                />
-                                <img
-                                    src={P2}
-                                    alt="Avatar 2"
-                                    className={`w-16 h-16 rounded-full cursor-pointer transform hover:scale-125 transition duration-300 ${avatar === P2 ? 'border-2 border-oraange' : ''}`}
-                                    onClick={() => handleAvatarSelect(P2)}
-                                />
-                                <img
-                                    src={P3}
-                                    alt="Avatar 3"
-                                    className={`w-16 h-16 rounded-full cursor-pointer transform hover:scale-125 transition duration-300 ${avatar === P3 ? 'border-2 border-oraange' : ''}`}
-                                    onClick={() => handleAvatarSelect(P3)}
-                                />
-                                <img
-                                    src={P7}
-                                    alt="Avatar 4"
-                                    className={`w-16 h-16 rounded-full cursor-pointer transform hover:scale-125 transition duration-300 ${avatar === P7 ? 'border-2 border-oraange' : ''}`}
-                                    onClick={() => handleAvatarSelect(P7)}
-                                />
-                                <img
-                                    src={P5}
-                                    alt="Avatar 5"
-                                    className={`w-16 h-16 rounded-full cursor-pointer transform hover:scale-125 transition duration-300 ${avatar === P5 ? 'border-2 border-oraange' : ''}`}
-                                    onClick={() => handleAvatarSelect(P5)}
-                                />
-                                <img
-                                    src={P6}
-                                    alt="Avatar 6"
-                                    className={`w-16 h-16 rounded-full cursor-pointer transform hover:scale-125 transition duration-300 ${avatar === P6 ? 'border-2 border-oraange' : ''}`}
-                                    onClick={() => handleAvatarSelect(P6)}
-                                />
-
-                            </div>
+                            <AvatarGallery handleAvatarSelect={handleAvatarSelect} avatar={avatar} />
                             <div className="w-full mb-4 mt-6">
                                 <label htmlFor="name" className="block font-bold text-lg">Username:</label>
                                 <input
